requirements.txt
# Core Data Processing
pandas>=2.0.0
duckdb>=0.9.0
openpyxl>=3.1.0  # For Excel file support

# Web Interface
streamlit>=1.28.0

# LLM Integration
openai>=1.0.0
google-generativeai>=0.3.0

# Configuration Management
python-dotenv>=1.0.0

# RAG System (Optional)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
